refactor: move complexity into leaf files where possible
